moved getImagesFromS3Bucket logic to ImageService,
